src/requirements.txt
python-multipart
fastapi
pydantic[email]
uvicorn
starlette
passlib[bcrypt]
python-jose[crytography]
sQlAlchemy
